requirements.txt
langsmith>=0.1.0
pandas>=1.5.0
python-dotenv>=1.0.0 